22 - Delete Permissions & Http Status Codes

+++ src/comments/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 from .forms import CommentForm
@@ -8,7 +8,21 @@ from .models import Comment
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    #obj = get_object_or_404(Comment, id=id)
+    # obj = Comment.objects.get(id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        #messages.success(request, "You do not have permission to view this.")
+        #raise Http404
+        reponse = HttpResponse("You do not have permission to do this.")
+        reponse.status_code = 403
+        return reponse
+        #return render(request, "confirm_delete.html", context, status_code=403)
+
     if request.method == "POST":
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -20,7 +34,15 @@ def comment_delete(request, id):
     return render(request, "confirm_delete.html", context)
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    #obj = Comment.objects.get(id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     content_object = obj.content_object # Post that the comment is on
     content_id = obj.content_object.id
